changed Timeline to Competencies
