+++ src/components/Header/HeaderCO.jsx
@@ -1,0 +1,37 @@
+import React from "react";
+import './Header.css';
+
+function Header() {
+
+    return(
+        <div>
+        <div className='inner-rectangle-8'>
+            <div class="marquee">
+                <ul class="marquee__content ">
+                    <li>&nbsp;Competencies&nbsp;</li>
+                    <li>&nbsp;Competencies&nbsp;</li>
+                    <li>&nbsp;Competencies&nbsp;</li>
+                    <li>&nbsp;Competencies&nbsp;</li>
+                    <li>&nbsp;Competencies&nbsp;</li>
+                    <li>&nbsp;Competencies&nbsp;</li>
+                    <li>&nbsp;Competencies&nbsp;</li>
+                </ul>
+                <ul class="marquee__content">
+                    <li>&nbsp;Competencies&nbsp;</li>
+                    <li>&nbsp;Competencies&nbsp;</li>
+                    <li>&nbsp;Competencies&nbsp;</li>
+                    <li>&nbsp;Competencies&nbsp;</li>
+                    <li>&nbsp;Competencies&nbsp;</li>
+                    <li>&nbsp;Competencies&nbsp;</li>
+                    <li>&nbsp;Competencies&nbsp;</li>
+                </ul>
+                </div>
+            </div>
+            <img src="https://live.staticflickr.com/65535/52783270220_87fdc16456_b.jpg" className='lionex-png'/>
+            <div className='line'>
+            </div>
+        </div>
+    );
+}
+
+export default Header
