+++ src/components/Navbar/Navbar.jsx
@@ -9,6 +9,7 @@ import projectIcon from '../../assets/Projects-icon.png'
 
 // import { IconName } from "react-icons/md";
 import { GrProjects } from "react-icons/gr";
+import { GiAchievement, GiSkills } from "react-icons/gi"
 
 function Navbar() {
 
@@ -19,22 +20,22 @@ return(
       <ul>
         <li>
           <Link to="/">
-            <FontAwesomeIcon icon={faHome}/>&nbsp; &nbsp; Overview
+            <FontAwesomeIcon icon={faHome}/>&nbsp; &nbsp;&nbsp;&nbsp;&nbsp; Overview
           </Link>
         </li>
         <li>
           <Link to="/about">
-            <FontAwesomeIcon icon={faInfoCircle}/>&nbsp; &nbsp; About
+            <FontAwesomeIcon icon={faInfoCircle}/>&nbsp; &nbsp;&nbsp;&nbsp;&nbsp; About
           </Link>
         </li>
         <li>
-          <Link to="/projects"><AiFillProject />&nbsp; &nbsp;Projects</Link>
+          <Link to="/projects"><AiFillProject />&nbsp; &nbsp;&nbsp;&nbsp;&nbsp;Projects</Link>
         </li>
         <li>
-          <Link to="/timeline"><FaClock />&nbsp; &nbsp; Timeline</Link>
+          <Link to="/competencies"><GiSkills/><GiAchievement />&nbsp;&nbsp;Competencies</Link>
         </li>
         <li>
-          <Link to="/contact"><AiOutlineMail/>&nbsp; &nbsp;  Contact</Link>
+          <Link to="/contact"><AiOutlineMail/>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;  Contact</Link>
         </li>
       </ul>
     </nav>
