+++ src/pages/4-Competencies/Competencies.css
